weather/requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3>=1.26.0
botocore>=1.29.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
deep-translator==1.11.4
Deprecated==1.2.14
docopt==0.6.2
Flask==3.0.3
Flask-Limiter==3.7.0
fonttools==4.53.1
h11==0.14.0
idna==3.7
importlib_resources==6.4.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
numpy==2.0.1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyvim==3.0.3
pyvmomi==8.0.3.0.1
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3>=1.25.4,<1.27
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
gunicorn==20.1.0
gevent>=1.4
prometheus_client==0.17.0
